move the common functions into utils.go and test_utils.go
