+++ cmd/build.go
@@ -2,7 +2,6 @@ package cmd
 
 import (
 	"bufio"
-	"bytes"
 	"fmt"
 	"io"
 	"os"
@@ -116,19 +115,6 @@ func builder(args []string, image string, file string) error {
 		return err
 	}
 	return nil
-}
-
-func execute(commandName string, params []string) (string, error) {
-	cmd := exec.Command(commandName, params...)
-	var out bytes.Buffer
-	cmd.Stdout = &out
-	cmd.Stderr = &out
-	err := cmd.Start()
-	if err != nil {
-		return "", err
-	}
-	err = cmd.Wait()
-	return out.String(), err
 }
 
 func printPipeOutput(pipe io.ReadCloser) {
