+++ cmd/test_utils.go
@@ -1,0 +1,22 @@
+
+package cmd
+
+import (
+	"bytes"
+	"os/exec"
+)
+
+// execute runs a command and returns the output and error
+
+func execute(commandName string, params []string) (string, error) {
+	cmd := exec.Command(commandName, params...)
+	var out bytes.Buffer
+	cmd.Stdout = &out
+	cmd.Stderr = &out
+	err := cmd.Start()
+	if err != nil {
+		return "", err
+	}
+	err = cmd.Wait()
+	return out.String(), err
+}
